added resources assert and re-org as per hive review feedback
